feat: new event page
